Fix the record keeping of indent levels
The problem was that the entire vector of levels always was serialized, while a
'0' always was added before deserializing started. This resulted in the vector
growing indefinitely and eventually serialization failed as the provided buffer
wouldn't be large enough.

The solution is, rather obviously, to pop the initial 0 before serialization.
That's not possible with a 'vector' though so the I changed to using a 'dequeu'
instead.

At the same time I also cleaned up some of the code, just minor things, but
worth doing.